Организует структуру кода
