+++ js/data.js
@@ -1,0 +1,40 @@
+//Создаем массив сообщений в комментарии
+const MESSAGES = [
+  'Всё отлично!',
+  ' В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+//создаем массив имен в комментарии
+const NAMES = [
+  'Дарья',
+  'Петя',
+  'Дмитрий',
+  'Никита',
+  'Ярослав',
+  'Светлана',
+  'Маша',
+  'Витя',
+  'Коля',
+  'Анастасия',
+  'Руслан',
+  'Кирилл',
+  'Наталья',
+  'Катя',
+  'Денис',
+  'Аня',
+  'Владимир',
+  'Аристарх',
+  'Данила',
+  'Женя',
+  'Юлия',
+  'Артём',
+  'Кристина',
+  'София',
+  'Лена',
+];
+
+export {MESSAGES, NAMES};
